refactored plog so it can be initialized from multiple places
