+++ standalone/cmake/CMakeLists_bgfx-android-arm64-v8a.txt
@@ -265,6 +265,8 @@ add_library(vpinball SHARED
    src/utils/hash.h
    src/utils/objloader.cpp
    src/utils/objloader.h
+   src/utils/logger.cpp
+   src/utils/logger.h
 
    third-party/include/imgui/imconfig.h
    third-party/include/imgui/imgui.cpp
