+++ standalone/cmake/CMakeLists_gl-android-arm64-v8a.txt
@@ -271,6 +271,8 @@ add_library(vpinball SHARED
    src/utils/hash.h
    src/utils/objloader.cpp
    src/utils/objloader.h
+   src/utils/logger.cpp
+   src/utils/logger.h
 
    third-party/include/imgui/imconfig.h
    third-party/include/imgui/imgui.cpp
